Prod corrections

+++ database/migrations/2020_08_04_093545_create_notes_tags.php
@@ -14,8 +14,8 @@ class CreateNotesTags extends Migration
     public function up()
     {
         Schema::create('notes_tags', function (Blueprint $table) {
-            $table->integer('note_id')->unsigned();
-            $table->integer('tag_id')->unsigned();
+            $table->bigInteger('note_id')->unsigned();
+            $table->bigInteger('tag_id')->unsigned();
             $table->foreign('note_id')->references('id')->on('notes')
                 ->onDelete('cascade');
             $table->foreign('tag_id')->references('id')->on('tags')
